Replaced helper class with model methods and a custom manager.
