+++ backend/workflow_manager/file_execution/models.py
@@ -1,4 +1,5 @@
 import uuid
+from typing import Optional
 
 from django.db import models
 from utils.models.base_model import BaseModel
@@ -9,6 +10,45 @@ FILE_NAME_LENGTH = 255
 FILE_PATH_LENGTH = 255
 HASH_LENGTH = 64
 MIME_TYPE_LENGTH = 128
+
+
+class WorkflowFileExecutionManager(models.Manager):
+    def get_or_create_file_execution(
+        self,
+        workflow_execution: WorkflowExecution,
+        file_name: str,
+        file_size: int,
+        file_hash: str,
+        file_path: Optional[str] = None,
+        mime_type: Optional[str] = None,
+    ):
+        """
+        Retrieves or creates a new input file record for a workflow execution.
+
+        Args:
+        workflow_execution: The `WorkflowExecution` object associated with this file
+        file_name: The name of the input file
+        file_size: The size of the file in bytes
+        file_hash: The hash of the file content
+        file_path: (Optional) The full path of the input file
+        mime_type: (Optional) MIME type of the file
+
+        return:
+            The `WorkflowFileExecution` object
+        """
+        execution_file: WorkflowFileExecution
+        execution_file, is_created = self.get_or_create(
+            workflow_execution=workflow_execution,
+            file_hash=file_hash,
+            file_path=file_path,
+        )
+
+        if is_created:
+            execution_file.file_name = file_name
+            execution_file.file_size = file_size
+            execution_file.mime_type = mime_type
+            execution_file.save()
+        return execution_file
 
 
 class WorkflowFileExecution(BaseModel):
@@ -49,11 +89,37 @@ class WorkflowFileExecution(BaseModel):
         blank=True, null=True, db_comment="Error message if execution failed"
     )
 
+    # Custom manager
+    objects = WorkflowFileExecutionManager()
+
     def __str__(self):
         return (
             f"WorkflowFileExecution: {self.file_name} "
             f"(WorkflowExecution: {self.workflow_execution})"
         )
+
+    def update_status(
+        self,
+        status: ExecutionStatus,
+        execution_time: int = 0,
+        execution_error: str = None,
+    ) -> None:
+        """
+        Updates the status and execution details of an input file.
+
+        Args:
+        execution_file: The `WorkflowExecutionFile` object to update
+        status: The new status of the file
+        execution_time: The execution time for processing the file
+        execution_error: (Optional) Error message if processing failed
+
+        return:
+            The updated `WorkflowExecutionInputFile` object
+        """
+        self.status = status
+        self.execution_time = execution_time
+        self.execution_error = execution_error
+        self.save()
 
     class Meta:
         verbose_name = "Workflow File Execution"
